# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateMicrotaskInput:
    properties:
      order:
        type: integer
      steps_planned:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
...
  dto.CreatePromiseRequest:
    properties:
      deadline:
        type: string
      description:
        maxLength: 2000
        type: string
      is_private:
        type: boolean
      title:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - deadline
    - title
    type: object
  dto.CreatePromiseWithMicrotasksRequest:
    properties:
      deadline:
        type: string
      description:
        type: string
      is_private:
        type: boolean
      microtasks:
        items:
          $ref: '#/definitions/dto.CreateMicrotaskInput'
        type: array
      title:
        type: string
    required:
    - deadline
    - title
    type: object
...
  dto.MicrotaskProgress:
    properties:
      completion_ratio:
        type: number
      id:
        type: string
      order:
        type: integer
      posts_count:
        type: integer
      status:
        type: string
      steps_planned:
        type: integer
      title:
        type: string
    type: object
...
  dto.PromiseResponse:
    properties:
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      status:
        type: string
      title:
        type: string
      username:
        description: можно добавить позже
        type: string
    type: object
  dto.PromiseWithMicrotasksProgressResponse:
    properties:
      deadline:
        type: string
      description:
        type: string
      id:
        type: string
      is_private:
        type: boolean
      microtasks:
        items:
          $ref: '#/definitions/dto.MicrotaskProgress'
        type: array
      status:
        type: string
      title:
        type: string
    type: object
...
  dto.UpdatePromiseRequest:
    properties:
      deadline:
        type: string
      description:
        maxLength: 2000
        type: string
      is_private:
        type: boolean
      title:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для отслеживания обещаний пользователей.
  title: iPromise API
  version: "1.0"
paths:
...
  /promises:
    post:
      consumes:
      - application/json
      description: Пользователь создаёт новое обещание
      parameters:
      - description: Данные обещания
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePromiseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Обещание создано'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверный формат данных'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Внутренняя ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать обещание
      tags:
      - promises
  /promises/{id}:
    delete:
      description: Удаляет обещание, если пользователь является владельцем
      parameters:
      - description: ID обещания
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Обещание удалено'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверный формат ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет прав'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить обещание
      tags:
      - promises
    get:
      description: Возвращает конкретное обещание (если оно публичное, пользователь
        — владелец или подписчик)
      parameters:
      - description: ID обещания (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PromiseResponse'
        "400":
          description: 'error: Некорректный ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет доступа'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Обещание не найдено'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить обещание по ID
      tags:
      - promises
    patch:
      consumes:
      - application/json
      description: Обновляет название, описание или дедлайн обещания
      parameters:
      - description: ID обещания
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePromiseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Обещание обновлено'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверные данные'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет доступа'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить обещание
      tags:
      - promises
...
  /promises/full:
    post:
      consumes:
      - application/json
      description: Создаёт новое обещание и вложенные микротаски (при необходимости)
        одним запросом
      parameters:
      - description: Данные обещания и микротасков
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePromiseWithMicrotasksRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Обещание создано'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверные данные'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать обещание с микротасками
      tags:
      - promises
  /promises/public:
    get:
      description: Возвращает все публичные обещания от всех пользователей
      parameters:
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Дата и время (RFC3339) для пагинации
        in: query
        name: after
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PromiseResponse'
            type: array
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Публичные обещания (все)
      tags:
      - promises
  /promises/user/{username}:
    get:
      description: Возвращает список обещаний в профиле пользователя (с учётом приватности)
      parameters:
      - description: Username пользователя
        in: path
        name: username
        required: true
        type: string
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339) для пагинации
        in: query
        name: after
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PromiseResponse'
            type: array
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Пользователь не найден'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обещания пользователя по username
      tags:
      - promises
  /promises/user/{username}/progress:
    get:
      description: Возвращает список всех обещаний указанного пользователя с вложенными
        микротасками и данными о прогрессе (кол-во постов, запланированные шаги, процент
        выполнения).
      parameters:
      - description: Username пользователя
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PromiseWithMicrotasksProgressResponse'
            type: array
        "400":
          description: 'error: Неверный username'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Пользователь не найден'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить все обещания пользователя с прогрессом по микротаскам
      tags:
      - promises
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"